+++ Modules/Bot.ts
@@ -1,17 +1,14 @@
 import { Client } from 'eris';
 import { config } from '../config';
-import { ErrorCache } from '../paths';
-import signale from 'signale';
 
 export const bot = new Client(config.token);
 
 bot.connect();
 
 bot.on('ready', () => {
-    signale.start('Bot ready.');
+    console.log('[startup] Bot ready.');
 });
 
 bot.on('error', (e: Error) => {
-    ErrorCache.add(e);
-    signale.error(e);
+    console.log(e);
 });
