+++ other/index.ts
@@ -1,4 +1,1 @@
-export * from './Bot';
-export * from './Superb';
 export * from './Points.schema';
-export * from './Points';
