--- paths/errors.ts
@@ -1,43 +1,0 @@
-import { Request, Response, Router } from 'express';
-import signale from 'signale';
-
-class ErrorPathClass {
-    public router: Router;
-    public errorCache: Error[];
-
-    constructor() {
-        this.router = Router();
-        this.errorCache = [];
-
-        this.router.get('/', this.handle.bind(this));
-    }
-
-    public add(err: Error) {
-        signale.error(err);
-
-        if (this.errorCache.length > 10) {
-            this.errorCache.pop();
-        }
-
-        return this.errorCache.push({ name: err.name, message: err.message });
-    }
-
-    /**
-     * Returns the error cache
-     *
-     * @private
-     * @param {Request} req
-     * @param {Response} res
-     * @memberof ErrorPathClass
-     */
-    private handle(req: Request, res: Response) {
-        if (this.errorCache.length === 0) {
-            return res.send('No errors.');
-        }
-        return res.send(this.errorCache);
-    }
-
-}
-
-export const ErrorCache = new ErrorPathClass();
-export const ErrorPath = ErrorCache.router;
