--- paths/index.ts
@@ -1,2 +1,0 @@
-export * from './upvote';
-export * from './errors';
